+++ plugin/twitter.js
@@ -1,14 +1,12 @@
-const Axios = require('axios');
-const mongodb = require('mongodb').MongoClient;
-const promisify = require('util').promisify;
+import {axiosProxied as axios} from '../utils/axiosProxied';
+import {MongoClient as mongodb} from "mongodb";
+import {promisify} from "util"
+import fs from "fs-extra";
 const exec = promisify(require('child_process').exec);
-const HttpsProxyAgent = require("https-proxy-agent");
-const fs = require('fs-extra');
 
-const CONFIG = require("../../config.json");
-const PROXY = CONFIG.proxy;
-const DB_PORT = 27017;
-const DB_PATH = "mongodb://127.0.0.1:" + DB_PORT;
+const CONFIG = global.config.twitter;
+const DB_PATH = global.config.mongoDB;
+const PERMISSION = CONFIG.permission;
 const BEARER_TOKEN = "Bearer AAAAAAAAAAAAAAAAAAAAANRILgAAAAAAnNwIzUejRCOuH5E6I8xnZz4puTs%3D1Zv7ttfk8LF81IUq16cHjhLTvJu4FA33AGWWjCpTnA";
 const MAX_SIZE = 4194304;
 const OPTION_MAP = {
@@ -33,7 +31,6 @@ const POSTTYPE_MAP = {
     "all" : [1, 1, 1, 1]
 } 
 
-let axios = false;
 let guest_token = "";
 let cookie = "";
 let connection = true;
@@ -63,17 +60,6 @@ async function checkConnection() {
     });
 }
 
-function setAgent() {
-    if (PROXY.startsWith("http")) {
-        axios = Axios.create({
-            proxy: false,
-            httpsAgent : new HttpsProxyAgent(PROXY)
-        });
-    }
-    else axios = Axios;
-}
-
-
 function opt_dict(post_option) {
     let [origin, retweet, reply, pic, cook] = POSTTYPE_MAP[post_option];
     return {
@@ -88,7 +74,7 @@ function opt_dict(post_option) {
 function toOptNl(option) {
     let {post} = option;
     let opt_string = "";
-    for (key in OPTION_MAP) {
+    for (let key in OPTION_MAP) {
         if (OPTION_MAP[key] == post) opt_string = key;
     }
     if (option.bbq == true) opt_string += "; 需要烤架";
@@ -99,7 +85,7 @@ function toOptNl(option) {
 function firstConnect() {
     checkConnection().then(res => {
         if (!res) {
-            console.error("Twitter无法连接，功能暂停");
+            console.error("Twitter无法连接，功能暂停\n", res);
         }
         else {
             getGuestToken();
@@ -113,12 +99,12 @@ function firstConnect() {
             }, 1*60*60*1000);
         }
     }).catch(err => {
-        console.error("Twitter无法连接，功能暂停");
+        console.error("Twitter无法连接，功能暂停\n", err);
     });
 }
 
 function httpHeader() {
-    return headers = {
+    return {
         "origin" : "https://twitter.com",
         "authorization" : BEARER_TOKEN,
         "cookie" : cookie,
@@ -363,7 +349,7 @@ function unSubscribe(name, context) {
                     let text = "";
                     if (result.value == null || !result.value.groups.includes(group_id)) {
                         console.error(result.value, group_id);
-                        replyFunc(context, "小火汁你压根就没订阅嗷", true);
+                        replyFunc(context, "并没有订阅这个人", true);
                         mongo.close();
                         return;
                     }
@@ -399,7 +385,8 @@ async function checkTwiTimeline() {
     let subscribes = await twitter_db.find({}).toArray();
     let check_interval = subscribes.length > 0 ? subscribes.length * 30 * 1000 : 5 * 60 * 1000;
     mongo.close();
-    
+    let i = 0;
+
     async function refreshTimeline() {
         await mongodb(DB_PATH, {useUnifiedTopology: true}).connect().then(async mongo => {
             const twitter_db = mongo.db('bot').collection('twitter');
@@ -596,7 +583,7 @@ async function format(tweet, end_point = false, context = false) {
         let pics = "";
         let src = "";
         if ("extended_entities" in tweet) {
-            for (entity in tweet.extended_entities) {
+            for (let entity in tweet.extended_entities) {
                 if (entity == "media") {
                     let media = tweet.extended_entities.media;
                     for (let i = 0; i < media.length; i++) {
@@ -604,24 +591,23 @@ async function format(tweet, end_point = false, context = false) {
                         if (media[i].type == "photo") {
                             src = [media[i].media_url_https.substring(0, media[i].media_url_https.length-4),
                                 `?format=${media[i].media_url_https.substring(media[i].media_url_https.length-3, media[i].media_url_https.length)}&name=4096x4096`].join("");
-                            pics += `[CQ:image,cache=0,file=${src}]`;
+                            let f = await global.download(src, true);
+                            pics += `[CQ:image,file=file:///${f}]`;
                         }
                         else if (media[i].type == "animated_gif") {
                             try {
-                                await exec(`ffmpeg -i ${media[i].video_info.variants[0].url} -loop 0 -y ${__dirname}/temp.gif`)
+                                let gifmp4 = await global.download(media[i].video_info.variants[0].url, true);
+                                let tmpgif = `${__dirname}/../tmp/temp.gif`;
+                                await exec(`ffmpeg -i "${gifmp4}" -loop 0 -y ${tmpgif}`)
                                     .then(async ({stdout, stderr}) => {
                                         if (stdout.length == 0) {
-                                            if (fs.statSync(`${__dirname}/temp.gif`).size < MAX_SIZE) {
-                                                let gif = fs.readFileSync(`${__dirname}/temp.gif`);
-                                                let base64gif = Buffer.from(gif, 'binary').toString('base64');
-                                                pics += `[CQ:image,file=base64://${base64gif}]`;
-                                            }
-                                            else pics += `这是一张动图 [CQ:image,cache=0,file=${media[i].media_url_https}]` + `动起来看这里${media[i].video_info.variants[0].url}`;
+                                            let gif = fs.readFileSync(tmpgif, {encoding: "base64"});
+                                            pics += `[CQ:image,file=base64://${gif}]`;
                                         }
-                                    })
+                                    });
                             } catch(err) {
                                 console.error(err);
-                                pics += `这是一张动图 [CQ:image,cache=0,file=${media[i].media_url_https}]` + `动起来看这里${media[i].video_info.variants[0].url}`;
+                                pics += `这是一张动图 [CQ:image,file=${media[i].media_url_https}]` + `动起来看这里${media[i].video_info.variants[0].url}`;
                             }
                         }
                         else if (media[i].type == "video") {
@@ -630,9 +616,10 @@ async function format(tweet, end_point = false, context = false) {
                                 if (media[i].video_info.variants[j].content_type == "video/mp4") mp4obj.push(media[i].video_info.variants[j]);
                             }
                             mp4obj.sort((a, b) => {return b.bitrate - a.bitrate;});
-                            payload.push(`[CQ:image,cache=0,file=${media[i].media_url_https}]`);
+                            payload.push(`[CQ:image,file=${media[i].media_url_https}]`);
                             if (context) {
-                                replyFunc(context, `[CQ:video,file=${mp4obj[0].url}]`);
+                                let f = await global.download(mp4obj[0].url, true);
+                                replyFunc(context, `[CQ:video,file=file:///${f}]`);
                             }
                             else payload.push(`视频地址: ${mp4obj[0].url}`);
                         }
@@ -663,7 +650,7 @@ async function format(tweet, end_point = false, context = false) {
                 let nchoice = parseInt(/\d/.exec(tweet.card.name)[0]);
                 let count = "";
                 let lable = "";
-                for (i = 1; i < nchoice + 1; i++) {
+                for (let i = 1; i < nchoice + 1; i++) {
                     lable = tweet.card.binding_values[`choice${i}_label`].string_value;
                     count = tweet.card.binding_values[`choice${i}_count`].string_value;
                     payload.push(`${lable}:  ${count}`);
@@ -785,6 +772,7 @@ async function addSub(name, option_nl, context) {
 }
 
 function twitterAggr(context) {
+    if (!CONFIG.enable) return false;
     if (connection && /^看看(.+?)的?((第[0-9]?[一二三四五六七八九]?条)|(上*条)|(最新))?\s?(推特|Twitter)$/i.test(context.message)) {	
 		let num = 1;
         let name = "";
@@ -816,6 +804,7 @@ function twitterAggr(context) {
         return true;
     }
     else if (connection && /^订阅(推特|Twitter)https:\/\/twitter.com\/.+(\/status\/\d+)?([>＞](.{2,}))?/i.test(context.message)) {
+        if (!global.permissionCheck(context, PERMISSION)) return true;
         let name = (/status\/\d+/.test(context.message) && /\.com\/(.+)\/status/.exec(context.message)[1] ||
                     /\.com\/(.+)[>＞]/.exec(context.message)[1]);
         let option_nl = /[>＞](?<option_nl>.{2,})/.exec(context.message)[1];
@@ -824,11 +813,13 @@ function twitterAggr(context) {
         return true;
     }
     else if (connection && /^订阅.+的?(推特|Twitter)([>＞](?<option_nl>.{2,}))?/i.test(context.message)) {
+        if (!global.permissionCheck(context, PERMISSION)) return true;
         let {groups : {name, option_nl}} = /订阅(?<name>.+)的?(推特|Twitter)([>＞](?<option_nl>.{2,}))?/i.exec(context.message);
         addSub(name, option_nl, context);
         return true;
     }
     else if (/^取消订阅.+的?(推特|Twitter)$/i.test(context.message)) {
+        if (!global.permissionCheck(context, PERMISSION)) return true;
         let name = /取消订阅(.+)的?(推特|Twitter)/i.exec(context.message)[1];
         unSubscribe(name, context);
         return true;
@@ -838,14 +829,12 @@ function twitterAggr(context) {
         return true;
     }
     else if (/^清空(推特|Twitter)订阅$/i.test(context.message)) {
-        if (/owner|admin/.test(context.sender.role)) clearSubs(context, context.group_id);
-        else replyFunc(context, '您配吗？');
+        if (global.permissionCheck(context, ["SU", "owner"])) clearSubs(context, context.group_id);
         return true;
     }
     else return false;
 }
 
-setAgent();
 firstConnect();
 
-module.exports = {twitterAggr, twitterReply, checkTwiTimeline, clearSubs};
+export default {twitterAggr, twitterReply, checkTwiTimeline, clearSubs};